Don't convert case for player email records
Deals with a unique constant violation error. devise_guests seems to
generate email addresses with capitals, but then when trying to find
the email it's converted to lower case which returns nothing, which then
in turn creates another Player record with the lower case email. Finally
the unique validation check seems to be case insensitive when modifying
Player records, so it fails. Ignoring the case for email in Devise
settings seems to be a remedy for this, but it will matter if and when
we want players to be able to save an account.