FG: Improve colour usage
Instead of randomly generating palettes, we now use a static set of 8
common flag colours. While this feels like a step back, it actually
produces far more 'flaggy' flags.

In addition to this, colours are now supplied randomly to each layer.
A side effect of this is that it's now possible for all layers to have
the same colour, and thus the flag can be a solid rectangle. This needs
to be fixed.